Setup basic framework
The basic structure off of which I'll build is setup. It should be
obvious to see how additional problems are added. There is still work to
be done in the frontend, however. I'd like to be able to say, in a brief
one line terminal call, "run these problems, and show me their answers".